+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/scheduler/job/versionActivity/ProductVersionActivityDTO.java
@@ -1,0 +1,13 @@
+package gov.healthit.chpl.scheduler.job.versionActivity;
+
+import gov.healthit.chpl.dto.ActivityDTO;
+import gov.healthit.chpl.dto.ProductVersionDTO;
+import lombok.Data;
+
+@Data
+public class ProductVersionActivityDTO extends ActivityDTO {
+    private static final long serialVersionUID = -4601289583619405361L;
+
+    private ProductVersionDTO originalVersion;
+    private ProductVersionDTO updatedVersion;
+}