+++ RELEASE_NOTES.md
@@ -1,5 +1,35 @@
 # Release Notes
 
+## Version 30.1.0
+_2 June 2021_
+
+### Features
+* Add cures statistics creation and basic email report
+* Update README documentation
+* Add developer data to version activity when creating a version
+* Incorporate listing status in certified product collection direct review counts
+
+### Bug Fixes
+* All activity metadata returned for Version and Listing activity
+
+---
+
+## Version 30.0.0
+_17 May 2021_
+
+### Breaking Changes
+* Remove background jobs, /data/job_types
+
+### Features
+* Add Developer, Product, and Version information to complaints
+* Move Surveillance Upload and Meaningful Use Upload background jobs to quartz scheduled jobs
+* Allow edit/view of Service Base URL List value
+
+### Flagged Features
+* Add listing-level fields to flexible upload
+
+---
+
 ## Version 29.3.0
 _4 May 2021_